+++ src/ui.c
@@ -1,211 +1,37 @@
-#include "editor.h"
 #include <ncurses.h>
-#include <stdlib.h>
 #include <ctype.h>
 #include <string.h>
-#include <dirent.h>
-#include <unistd.h>
-#include <sys/stat.h>
-#include <stdio.h>
-#include <strings.h>
 #include "config.h"
 #include "ui.h"
 
-// Ensure the correct prototypes are available
-char *strdup(const char *s);
-
-static WINDOW *create_centered_window(int height, int width, WINDOW *parent) {
-    int win_y, win_x;
-
-    if (parent) {
-        int py, px, ph, pw;
-        getbegyx(parent, py, px);
-        getmaxyx(parent, ph, pw);
-        win_y = py + (ph - height) / 2;
-        win_x = px + (pw - width) / 2;
-    } else {
-        win_y = (LINES - height) / 2;
-        win_x = (COLS - width) / 2;
-    }
-
-    WINDOW *win = newwin(height, width, win_y, win_x);
-    if (!win) {
-        endwin();
-        fprintf(stderr, "Failed to create new window\n");
-        exit(EXIT_FAILURE);
-    }
-
-    box(win, 0, 0);
-
-    return win;
-}
-
-WINDOW *create_popup_window(int height, int width, WINDOW *parent) {
-    return create_centered_window(height, width, parent);
-}
-
-/**
- * Displays the help dialog with a list of available commands and their corresponding keybindings.
- */
-void show_help() {
-    // Window size and position adjustments
-    int win_height = 18;
-    int win_width = COLS - 40;
-    int win_y = (LINES - win_height) / 2;
-    int win_x = (COLS - win_width) / 2;
-
-    // Create a temporary window for the help dialog
-    WINDOW *help_win = newwin(win_height, win_width, win_y, win_x);
-    keypad(help_win, TRUE);  // Enable keyboard input for the help dialog
-
-    // Set color pair for the help window
-    wbkgd(help_win, COLOR_PAIR(1));  // Use the defined color pair
-    wrefresh(stdscr);  // Refresh the main screen before drawing the dialog
-
-    // Draw the help dialog borders
-    box(help_win, 0, 0);
-
-    // Print the help information in two columns
-
-    // Print the title
-    mvwprintw(help_win, 1, 2, "Help:");
-
-    // Print the command and keybinding for each command
-    mvwprintw(help_win, 3, 2, "CTRL-H: Show this help");
-    mvwprintw(help_win, 4, 2, "CTRL-A: About");
-    mvwprintw(help_win, 5, 2, "CTRL-L: Load a new file (browse files)");
-    mvwprintw(help_win, 6, 2, "CTRL-O: Save As (browse dir or name)");
-    mvwprintw(help_win, 7, 2, "CTRL-S: Save");
-    mvwprintw(help_win, 8, 2, "CTRL-J: Start/Stop selection mode");
-    mvwprintw(help_win, 9, 2, "CTRL-V: Paste from clipboard");
-    mvwprintw(help_win, 10, 2, "CTRL-N: New file");
-    mvwprintw(help_win, 11, 2, "CTRL-Y: Redo");
-    mvwprintw(help_win, 12, 2, "CTRL-Z: Undo");
-    mvwprintw(help_win, 13, 2, "CTRL-F: Search for text string");
-    mvwprintw(help_win, 14, 2, "F3: Find next occurrence");
-    mvwprintw(help_win, 15, 2, "CTRL-R: Replace");
-
-    // Print the second column of commands and keybindings
-    mvwprintw(help_win, 3, win_width / 2, "CTRL-X: Quit");
-    mvwprintw(help_win, 4, win_width / 2, "CTRL-W: Move forward to next word");
-    mvwprintw(help_win, 5, win_width / 2, "CTRL-B: Move backward to previous word");
-    mvwprintw(help_win, 6, win_width / 2, "CTRL-D: Delete current line");
-    mvwprintw(help_win, 7, win_width / 2, "Arrow Keys: Navigate text");
-    mvwprintw(help_win, 8, win_width / 2, "Page Up/Down: Scroll document");
-    mvwprintw(help_win, 9, win_width / 2, "CTRL-Left: Move to line start");
-    mvwprintw(help_win, 10, win_width / 2, "CTRL-Right: Move to line end");
-    mvwprintw(help_win, 11, win_width / 2, "CTRL-PgUp: Move to top of doc");
-    mvwprintw(help_win, 12, win_width / 2, "CTRL-PgDn: Move to end of doc");
-    mvwprintw(help_win, 14, win_width / 2, "F5: Insert blank line");
-    mvwprintw(help_win, 15, win_width / 2, "F6: Next file");
-    mvwprintw(help_win, 16, win_width / 2, "F7: Previous file");
-
-    // Wait for any keypress to close the dialog
-    mvwprintw(help_win, win_height - 1, (win_width - strlen("(Press any key to close)")) / 2, "(Press any key to close)");
-    wrefresh(help_win);
-    wgetch(help_win);
-
-    // Clean up the help window
-    wclear(help_win);
-    wrefresh(help_win);
-    delwin(help_win);
-    wrefresh(stdscr);  // Refresh the main screen after closing the dialog
-}
-
-/**
- * Displays the about dialog with information about the Vento Text Editor.
- */
-void show_about() {
-    // Window size and position adjustments
-    int win_height = 10;
-    int win_width = COLS - 20;
-    int win_y = (LINES - win_height) / 2;
-    int win_x = (COLS - win_width) / 2;
-
-    // Create a temporary window for the about dialog
-    WINDOW *about_win = newwin(win_height, win_width, win_y, win_x);
-    keypad(about_win, TRUE);  // Enable keyboard input for the about dialog
-
-    // Set color pair for the about window
-    wbkgd(about_win, COLOR_PAIR(1));  // Use the defined color pair
-    wrefresh(stdscr);  // Refresh the main screen before drawing the dialog
-
-    // Draw the about dialog borders
-    box(about_win, 0, 0);
-
-    // Print the about information
-
-    // Print the title
-    mvwprintw(about_win, 1, 2, "Vento Text Editor");
-
-    // Print the version
-    mvwprintw(about_win, 2, 2, "Version: %s", VERSION);
-
-    // Print the license
-    mvwprintw(about_win, 3, 2, "License: GPL v3");
-
-    // Print the description
-    mvwprintw(about_win, 4, 2, "Vento is open-source software licensed under the GPL v3.");
-
-    // Wait for any keypress to close the dialog
-    mvwprintw(about_win, win_height - 2, 2, "(Press any key to close)");
-    wrefresh(about_win);
-    wgetch(about_win);
-
-    // Clean up the about window
-    wclear(about_win);
-    wrefresh(about_win);
-    delwin(about_win);
-    wrefresh(stdscr);  // Refresh the main screen after closing the dialog
-}
-
-/**
- * Creates a dialog window with a message and an input field for the user to enter text.
- * The entered text is stored in the 'output' parameter.
- * The maximum length of the input is specified by 'max_input_len'.  Press
- * Enter to accept the input or press Esc to cancel the dialog.  When
- * cancelled, the output string is cleared.
- *
- * @param message The message to display in the dialog window.
- * @param output  The buffer to store the user's input.
- * @param max_input_len The maximum length of the user's input.
- */
 void create_dialog(const char *message, char *output, int max_input_len) {
-    // Define colors
     start_color();
-    init_pair(1, COLOR_WHITE, COLOR_BLUE); // Background color for dialog
-    init_pair(2, COLOR_YELLOW, COLOR_BLACK); // Input text color
+    init_pair(1, COLOR_WHITE, COLOR_BLUE);
+    init_pair(2, COLOR_YELLOW, COLOR_BLACK);
 
-    // Window size and position adjustments
     int win_y = LINES / 3;
     int win_x = (COLS - strlen(message) - 30) / 2;
     int win_width = strlen(message) + 30;
     int win_height = 7;
 
-    // Create a temporary window for the dialog
     WINDOW *dialog_win = newwin(win_height, win_width, win_y, win_x);
-    keypad(dialog_win, TRUE);  // Enable keyboard input for the dialog
-    wbkgd(dialog_win, COLOR_PAIR(1)); // Set background color
-    wrefresh(stdscr);  // Refresh the main screen before drawing the dialog
+    keypad(dialog_win, TRUE);
+    wbkgd(dialog_win, COLOR_PAIR(1));
+    wrefresh(stdscr);
 
-    // Draw the dialog borders
     box(dialog_win, 0, 0);
 
-    // Print the message at the center
-    wattron(dialog_win, A_BOLD); // Bold the message
+    wattron(dialog_win, A_BOLD);
     mvwprintw(dialog_win, 1, (win_width - strlen(message)) / 2, "%s", message);
     wattroff(dialog_win, A_BOLD);
 
-    // Input field and cursor position
-    int input_x = 2; // Adjust as needed for message positioning
+    int input_x = 2;
     int input_y = 3;
-    wattron(dialog_win, COLOR_PAIR(2)); // Set input text color
+    wattron(dialog_win, COLOR_PAIR(2));
     mvwprintw(dialog_win, input_y, input_x, "Input: ");
     wattroff(dialog_win, COLOR_PAIR(2));
-    wmove(dialog_win, input_y, input_x + 7);  // Move cursor after "Input: "
+    wmove(dialog_win, input_y, input_x + 7);
 
-    // Get user input with maximum length check.  Pressing Esc cancels the dialog
-    // and clears the output string.
     int ch, input_len = 0;
     int cancelled = 0;
     while ((ch = wgetch(dialog_win)) != '\n' && input_len < max_input_len - 1) {
@@ -222,7 +48,7 @@ void create_dialog(const char *message, char *output, int max_input_len) {
                 wmove(dialog_win, input_y, input_x + 7 + input_len);
                 wrefresh(dialog_win);
             }
-        } else if (ch == 27) { // ESC key cancels the dialog
+        } else if (ch == 27) {
             cancelled = 1;
             break;
         }
@@ -230,532 +56,39 @@ void create_dialog(const char *message, char *output, int max_input_len) {
     if (cancelled)
         output[0] = '\0';
     else
-        output[input_len] = '\0';  // Add null terminator to the input string
+        output[input_len] = '\0';
 
-    // Clean up the dialog window
     wclear(dialog_win);
     wrefresh(dialog_win);
     delwin(dialog_win);
-    wrefresh(stdscr);  // Refresh the main screen after closing the dialog
-
-}
-
-/**
- * Displays a warning dialog to inform the user that the software is experimental and not intended for production use.
- * The dialog will be dismissed when any key is pressed.
- */
-void show_warning_dialog() {
-    int win_height = 7;
-    int win_width = COLS - 20;
-    int win_y = (LINES - win_height) / 2;
-    int win_x = (COLS - win_width) / 2;
-
-    // Create a temporary window for the warning dialog
-    WINDOW *warning_win = newwin(win_height, win_width, win_y, win_x);
-    keypad(warning_win, TRUE);  // Enable keyboard input for the warning dialog
-
-    // Set color pair for the warning window
-    wbkgd(warning_win, COLOR_PAIR(1));
-    wrefresh(stdscr);  // Refresh the main screen before drawing the dialog
-
-    // Draw the warning dialog borders
-    box(warning_win, 0, 0);
-
-    // Print the warning message centered
-    char *message1 = "Warning: This is experimental software.";
-    char *message2 = "It is under development and not intended for production use.";
-    char *message3 = "(Press any key to dismiss)";
-
-    mvwprintw(warning_win, 2, (win_width - strlen(message1)) / 2, "%s", message1);
-    mvwprintw(warning_win, 3, (win_width - strlen(message2)) / 2, "%s", message2);
-    mvwprintw(warning_win, 5, (win_width - strlen(message3)) / 2, "%s", message3);
-
-    wrefresh(warning_win);
-
-    // Wait for any keypress to close the dialog
-    wgetch(warning_win);
-
-    // Refresh the entire screen
-    werase(text_win);
-    box(text_win, 0, 0);
-    draw_text_buffer(active_file, text_win);
-    wrefresh(text_win);
-    update_status_bar(active_file);
-    wrefresh(stdscr);  // Refresh the main screen after closing the dialog
-}
-
-void get_dir_contents(const char *dir_path, char ***choices, int *n_choices) {
-    DIR *dir;
-    struct dirent *entry;
-    int count = 0;
-
-    dir = opendir(dir_path);
-    if (!dir) {
-        *choices = NULL;
-        *n_choices = 0;
-        return;
-    }
-
-    // Count entries
-    while ((entry = readdir(dir)) != NULL) {
-        count++;
-    }
-    rewinddir(dir);
-
-    *choices = (char **)malloc(count * sizeof(char *));
-    if (!*choices) {
-        closedir(dir);
-        *choices = NULL;
-        *n_choices = 0;
-        return;
-    }
-
-    int i = 0;
-    while ((entry = readdir(dir)) != NULL) {
-        (*choices)[i] = strdup(entry->d_name);
-        if (!(*choices)[i]) {
-            for (int j = 0; j < i; ++j) {
-                free((*choices)[j]);
-            }
-            free(*choices);
-            closedir(dir);
-            *choices = NULL;
-            *n_choices = 0;
-            return;
-        }
-        i++;
-    }
-    closedir(dir);
-
-    *n_choices = count;
-}
-
-void free_dir_contents(char **choices, int n_choices) {
-    for (int i = 0; i < n_choices; ++i) {
-        free(choices[i]);
-    }
-    free(choices);
-}
-
-
-int show_open_file_dialog(char *path, int max_len) {
-    int highlight = 0;
-    int ch;
-    char cwd[1024];
-    char input[1024] = "";
-    int input_len = 0;
-    int start = 0;
-
-    int win_height = LINES - 4;
-    int win_width = COLS - 4;
-    WINDOW *win = create_popup_window(win_height, win_width, NULL);
-    if (!win)
-        return 0;
-    keypad(win, TRUE);
-
-    getcwd(cwd, sizeof(cwd));
-
-    while (1) {
-        werase(win);
-        box(win, 0, 0);
-        mvwprintw(win, 1, 2, "Current Directory: %s", cwd);
-
-        char **choices = NULL;
-        int n_choices = 0;
-        get_dir_contents(cwd, &choices, &n_choices);
-
-        int max_display = win_height - 5; // reserve lines for dir, instructions and input
-        if (highlight < start)
-            start = highlight;
-        if (highlight >= start + max_display)
-            start = highlight - max_display + 1;
-
-        for (int i = 0; i < max_display && i + start < n_choices; ++i) {
-            int idx = i + start;
-            if (idx == highlight)
-                wattron(win, A_REVERSE);
-            mvwprintw(win, i + 2, 2, "%s", choices[idx]);
-            wattroff(win, A_REVERSE);
-        }
-
-        for (int i = n_choices - start; i < max_display; ++i) {
-            mvwprintw(win, i + 2, 2, "%*s", win_width - 4, "");
-        }
-
-        mvwprintw(win, win_height - 3, 2, "Arrows: move  Enter: select  ESC: cancel");
-        mvwprintw(win, win_height - 2, 2, "Path: %s", input);
-        wmove(win, win_height - 2, 8 + input_len);
-        wrefresh(win);
-
-        ch = wgetch(win);
-        if (ch == KEY_UP) {
-            if (highlight > 0)
-                --highlight;
-        } else if (ch == KEY_DOWN) {
-            if (highlight < n_choices - 1)
-                ++highlight;
-        } else if (ch == '\n') {
-            if (input_len > 0) {
-                char tmp[2048];
-                if (input[0] == '/') {
-                    strncpy(tmp, input, sizeof(tmp) - 1);
-                    tmp[sizeof(tmp) - 1] = '\0';
-                } else {
-                    snprintf(tmp, sizeof(tmp), "%s/%s", cwd, input);
-                }
-                strncpy(path, tmp, max_len);
-                path[max_len - 1] = '\0';
-                free_dir_contents(choices, n_choices);
-                wclear(win);
-                wrefresh(win);
-                delwin(win);
-                wrefresh(stdscr);
-                return 1;
-            }
-
-            if (n_choices > 0) {
-                struct stat sb;
-                char next_path[2048];
-                snprintf(next_path, sizeof(next_path), "%s/%s", cwd, choices[highlight]);
-                if (stat(next_path, &sb) == 0 && S_ISDIR(sb.st_mode)) {
-                    strncpy(cwd, next_path, sizeof(cwd));
-                    cwd[sizeof(cwd) - 1] = '\0';
-                    highlight = 0;
-                    start = 0;
-                    input_len = 0;
-                    input[0] = '\0';
-                } else {
-                    strncpy(path, next_path, max_len);
-                    path[max_len - 1] = '\0';
-                    free_dir_contents(choices, n_choices);
-                    wclear(win);
-                    wrefresh(win);
-                    delwin(win);
-                    wrefresh(stdscr);
-                    return 1;
-                }
-            }
-        } else if (ch == KEY_BACKSPACE || ch == 127) {
-            if (input_len > 0) {
-                input_len--;
-                input[input_len] = '\0';
-            }
-        } else if (ch == 27) {
-            free_dir_contents(choices, n_choices);
-            wclear(win);
-            wrefresh(win);
-            delwin(win);
-            wrefresh(stdscr);
-            return 0;
-        } else if (isprint(ch)) {
-            if (input_len < (int)sizeof(input) - 1) {
-                input[input_len++] = ch;
-                input[input_len] = '\0';
-            }
-        }
-
-        free_dir_contents(choices, n_choices);
-    }
-
-    wclear(win);
-    wrefresh(win);
-    delwin(win);
     wrefresh(stdscr);
-    return 0;
 }
 
-int show_save_file_dialog(char *path, int max_len) {
-    int highlight = 0;
-    int ch;
-    char cwd[1024];
-    char input[1024] = "";
-    int input_len = 0;
-    int start = 0;
-
-    int win_height = LINES - 4;
-    int win_width = COLS - 4;
-    WINDOW *win = create_popup_window(win_height, win_width, NULL);
-    if (!win)
-        return 0;
-    keypad(win, TRUE);
-
-    getcwd(cwd, sizeof(cwd));
-
-    while (1) {
-        werase(win);
-        box(win, 0, 0);
-        mvwprintw(win, 1, 2, "Current Directory: %s", cwd);
-
-        char **choices = NULL;
-        int n_choices = 0;
-        get_dir_contents(cwd, &choices, &n_choices);
-
-        int max_display = win_height - 5; // reserve lines for dir, instructions and input
-        if (highlight < start)
-            start = highlight;
-        if (highlight >= start + max_display)
-            start = highlight - max_display + 1;
-
-        for (int i = 0; i < max_display && i + start < n_choices; ++i) {
-            int idx = i + start;
-            if (idx == highlight)
-                wattron(win, A_REVERSE);
-            mvwprintw(win, i + 2, 2, "%s", choices[idx]);
-            wattroff(win, A_REVERSE);
-        }
-
-        for (int i = n_choices - start; i < max_display; ++i) {
-            mvwprintw(win, i + 2, 2, "%*s", win_width - 4, "");
-        }
-
-        mvwprintw(win, win_height - 3, 2, "Arrows: move  Enter: select  ESC: cancel");
-        mvwprintw(win, win_height - 2, 2, "Path: %s", input);
-        wmove(win, win_height - 2, 8 + input_len);
-        wrefresh(win);
-
-        ch = wgetch(win);
-        if (ch == KEY_UP) {
-            if (highlight > 0)
-                --highlight;
-        } else if (ch == KEY_DOWN) {
-            if (highlight < n_choices - 1)
-                ++highlight;
-        } else if (ch == '\n') {
-            if (input_len > 0) {
-                char tmp[2048];
-                if (input[0] == '/') {
-                    strncpy(tmp, input, sizeof(tmp) - 1);
-                    tmp[sizeof(tmp) - 1] = '\0';
-                } else {
-                    snprintf(tmp, sizeof(tmp), "%s/%s", cwd, input);
-                }
-                strncpy(path, tmp, max_len);
-                path[max_len - 1] = '\0';
-                free_dir_contents(choices, n_choices);
-                wclear(win);
-                wrefresh(win);
-                delwin(win);
-                wrefresh(stdscr);
-                return 1;
-            }
-
-            if (n_choices > 0) {
-                struct stat sb;
-                char next_path[2048];
-                snprintf(next_path, sizeof(next_path), "%s/%s", cwd, choices[highlight]);
-                if (stat(next_path, &sb) == 0 && S_ISDIR(sb.st_mode)) {
-                    strncpy(cwd, next_path, sizeof(cwd));
-                    cwd[sizeof(cwd) - 1] = '\0';
-                    highlight = 0;
-                    start = 0;
-                    input_len = 0;
-                    input[0] = '\0';
-                } else {
-                    strncpy(path, next_path, max_len);
-                    path[max_len - 1] = '\0';
-                    free_dir_contents(choices, n_choices);
-                    wclear(win);
-                    wrefresh(win);
-                    delwin(win);
-                    wrefresh(stdscr);
-                    return 1;
-                }
-            }
-        } else if (ch == KEY_BACKSPACE || ch == 127) {
-            if (input_len > 0) {
-                input_len--;
-                input[input_len] = '\0';
-            }
-        } else if (ch == 27) {
-            free_dir_contents(choices, n_choices);
-            wclear(win);
-            wrefresh(win);
-            delwin(win);
-            wrefresh(stdscr);
-            return 0;
-        } else if (isprint(ch)) {
-            if (input_len < (int)sizeof(input) - 1) {
-                input[input_len++] = ch;
-                input[input_len] = '\0';
-            }
-        }
-
-        free_dir_contents(choices, n_choices);
-    }
-
-    wclear(win);
-    wrefresh(win);
-    delwin(win);
-    wrefresh(stdscr);
-    return 0;
-}
-
-int show_scrollable_window(const char **options, int count, WINDOW *parent) {
-    int highlight = 0;
-    int ch;
-    int start = 0;
-
-    if (count <= 0)
-        return -1;
-
-    int own = 0;
-    int win_height, win_width;
-    WINDOW *win;
-    if (parent) {
-        int ph, pw;
-        getmaxyx(parent, ph, pw);
-        win_height = ph - 4;
-        win_width = pw - 4;
-        win = create_popup_window(win_height, win_width, parent);
-        if (!win)
-            return -1;
-        own = 1;
-    } else {
-        win_height = LINES - 4;
-        win_width = COLS - 4;
-        win = create_popup_window(win_height, win_width, NULL);
-        if (!win)
-            return -1;
-        own = 1;
-    }
-    keypad(win, TRUE);
-
-    while (1) {
-        werase(win);
-        box(win, 0, 0);
-
-        int max_display = win_height - 2; // leave last line for instructions
-        if (highlight < start)
-            start = highlight;
-        if (highlight >= start + max_display)
-            start = highlight - max_display + 1;
-
-        for (int i = 0; i < max_display && i + start < count; ++i) {
-            int idx = i + start;
-            if (idx == highlight)
-                wattron(win, A_REVERSE);
-            mvwprintw(win, i + 1, 1, "%s", options[idx]);
-            wattroff(win, A_REVERSE);
-        }
-
-        for (int i = count - start; i < max_display; ++i) {
-            mvwprintw(win, i + 1, 1, "%*s", win_width - 2, "");
-        }
-
-        mvwprintw(win, win_height - 1, 1, "Arrows: move  Enter: select  ESC: cancel");
-        wrefresh(win);
-
-        ch = wgetch(win);
-        if (ch == KEY_UP) {
-            if (highlight > 0)
-                --highlight;
-        } else if (ch == KEY_DOWN) {
-            if (highlight < count - 1)
-                ++highlight;
-        } else if (ch == '\n') {
-            if (own) {
-                werase(win);
-                wrefresh(win);
-                delwin(win);
-                if (parent) {
-                    touchwin(parent);
-                    wrefresh(parent);
-                } else {
-                    wrefresh(stdscr);
-                }
-            }
-            return highlight;
-        } else if (ch == 27) {
-            if (own) {
-                werase(win);
-                wrefresh(win);
-                delwin(win);
-                if (parent) {
-                    touchwin(parent);
-                    wrefresh(parent);
-                } else {
-                    wrefresh(stdscr);
-                }
-            }
-            return -1;
-        } else if (ch == KEY_MOUSE && enable_mouse) {
-            MEVENT ev;
-            if (getmouse(&ev) == OK &&
-                (ev.bstate & (BUTTON1_PRESSED | BUTTON1_CLICKED |
-                               BUTTON1_RELEASED))) {
-                int wy, wx;
-                getbegyx(win, wy, wx);
-                int row = ev.y - wy - 1;
-                int col = ev.x - wx - 1;
-                int max_display = win_height - 2;
-                if (row >= 0 && row < max_display &&
-                    col >= 0 && col < win_width - 2) {
-                    int idx = start + row;
-                    if (idx >= 0 && idx < count)
-                        highlight = idx;
-
-                    if (ev.bstate & (BUTTON1_RELEASED | BUTTON1_CLICKED)) {
-                        if (own) {
-                            werase(win);
-                            wrefresh(win);
-                            delwin(win);
-                            if (parent) {
-                                touchwin(parent);
-                                wrefresh(parent);
-                            } else {
-                                wrefresh(stdscr);
-                            }
-                        }
-                        return highlight;
-                    }
-                }
-            }
-        }
-    }
-
-    return -1;
-}
-
-/**
- * Shows a find dialog window and accepts a string from the user.
- * The entered string is stored in the 'output' parameter.
- * The maximum length of the input is specified by 'max_input_len'.
- *
- * @param output The buffer to store the user's input.
- * @param max_input_len The maximum length of the user's input.
- * @param preset Initial value to pre-fill the input field.
- */
 int show_find_dialog(char *output, int max_input_len, const char *preset) {
-    // Define colors
     start_color();
-    init_pair(1, COLOR_WHITE, COLOR_BLUE); // Background color for dialog
-    init_pair(2, COLOR_YELLOW, COLOR_BLACK); // Input text color
+    init_pair(1, COLOR_WHITE, COLOR_BLUE);
+    init_pair(2, COLOR_YELLOW, COLOR_BLACK);
 
-    // Window size and position adjustments
     int win_y = LINES / 3;
     int win_x = (COLS - 40) / 2;
     int win_width = 40;
     int win_height = 7;
 
-    // Create a temporary window for the dialog
     WINDOW *dialog_win = newwin(win_height, win_width, win_y, win_x);
-    keypad(dialog_win, TRUE);  // Enable keyboard input for the dialog
-    wbkgd(dialog_win, COLOR_PAIR(1)); // Set background color
-    wrefresh(stdscr);  // Refresh the main screen before drawing the dialog
+    keypad(dialog_win, TRUE);
+    wbkgd(dialog_win, COLOR_PAIR(1));
+    wrefresh(stdscr);
 
-    // Draw the dialog borders
     box(dialog_win, 0, 0);
 
-    // Print the message at the center
     char *message = "Find: ";
-    wattron(dialog_win, A_BOLD); // Bold the message
+    wattron(dialog_win, A_BOLD);
     mvwprintw(dialog_win, 1, (win_width - strlen(message)) / 2, "%s", message);
     wattroff(dialog_win, A_BOLD);
 
-    // Input field and cursor position
-    int input_x = 7; // Adjust as needed for message positioning
+    int input_x = 7;
     int input_y = 3;
-    wattron(dialog_win, COLOR_PAIR(2)); // Set input text color
+    wattron(dialog_win, COLOR_PAIR(2));
     mvwprintw(dialog_win, input_y, input_x, "Input: ");
     wattroff(dialog_win, COLOR_PAIR(2));
     int input_len = 0;
@@ -767,8 +100,7 @@ int show_find_dialog(char *output, int max_input_len, const char *preset) {
     } else {
         output[0] = '\0';
     }
-    wmove(dialog_win, input_y, input_x + 7 + input_len);  // Move cursor after "Input: "
-    // Get user input with maximum length check
+    wmove(dialog_win, input_y, input_x + 7 + input_len);
     int ch;
     int cancelled = 0;
     while ((ch = wgetch(dialog_win)) != '\n' && input_len < max_input_len - 1) {
@@ -789,7 +121,7 @@ int show_find_dialog(char *output, int max_input_len, const char *preset) {
                     wmove(dialog_win, input_y, input_x + 7 + input_len);
                     wrefresh(dialog_win);
                 }
-            } else if (ch == 27) { // ESC key
+            } else if (ch == 27) {
                 cancelled = 1;
                 break;
             }
@@ -800,13 +132,12 @@ int show_find_dialog(char *output, int max_input_len, const char *preset) {
     if (cancelled)
         output[0] = '\0';
     else
-        output[input_len] = '\0';  // Add null terminator
+        output[input_len] = '\0';
 
-    // Clean up the dialog window
     wclear(dialog_win);
     wrefresh(dialog_win);
     delwin(dialog_win);
-    wrefresh(stdscr);  // Refresh the main screen after closing the dialog
+    wrefresh(stdscr);
 
     return cancelled ? 0 : 1;
 }
@@ -819,422 +150,3 @@ int show_replace_dialog(char *search, int max_search_len,
     create_dialog("Replace:", replace, max_replace_len);
     return 1;
 }
-
-static void show_message(const char *msg) {
-    int win_height = 3;
-    int win_width = strlen(msg) + 4;
-    int win_y = (LINES - win_height) / 2;
-    int win_x = (COLS - win_width) / 2;
-
-    WINDOW *win = newwin(win_height, win_width, win_y, win_x);
-    box(win, 0, 0);
-    mvwprintw(win, 1, 2, "%s", msg);
-    wrefresh(win);
-    wgetch(win);
-    wclear(win);
-    wrefresh(win);
-    delwin(win);
-    wrefresh(stdscr);
-}
-
-static void str_to_upper(char *dst, const char *src, size_t dst_size) {
-    size_t i;
-    for (i = 0; i + 1 < dst_size && src[i]; ++i) {
-        dst[i] = toupper((unsigned char)src[i]);
-    }
-    dst[i] = '\0';
-}
-
-int show_settings_dialog(AppConfig *cfg) {
-    AppConfig original = *cfg;
-
-    enum {
-        FIELD_ENABLE_COLOR,
-        FIELD_MOUSE,
-        FIELD_BACKGROUND,
-        FIELD_KEYWORD,
-        FIELD_COMMENT,
-        FIELD_STRING,
-        FIELD_TYPE,
-        FIELD_SYMBOL,
-        FIELD_COUNT
-    };
-
-    const char *labels[FIELD_COUNT] = {
-        "Enable color",
-        "Enable mouse",
-        "Background color",
-        "Keyword color",
-        "Comment color",
-        "String color",
-        "Type color",
-        "Symbol color"
-    };
-
-    int highlight = 0;
-    int ch;
-    int done = 0;
-
-    int win_height = FIELD_COUNT + 4;
-    int longest = 0;
-    for (int i = 0; i < FIELD_COUNT; ++i) {
-        const char *val = "";
-        switch (i) {
-        case FIELD_ENABLE_COLOR:
-            val = cfg->enable_color ? "Enabled" : "Disabled";
-            break;
-        case FIELD_MOUSE:
-            val = cfg->enable_mouse ? "Enabled" : "Disabled";
-            break;
-        case FIELD_BACKGROUND:
-            val = cfg->background_color;
-            break;
-        case FIELD_KEYWORD:
-            val = cfg->keyword_color;
-            break;
-        case FIELD_COMMENT:
-            val = cfg->comment_color;
-            break;
-        case FIELD_STRING:
-            val = cfg->string_color;
-            break;
-        case FIELD_TYPE:
-            val = cfg->type_color;
-            break;
-        case FIELD_SYMBOL:
-            val = cfg->symbol_color;
-            break;
-        }
-        int len = strlen(labels[i]) + 2 + strlen(val);
-        if (len > longest)
-            longest = len;
-    }
-    int win_width = longest + 4;
-    if (win_width < 50)
-        win_width = 50;
-
-    WINDOW *win = create_popup_window(win_height, win_width, NULL);
-    if (!win)
-        return 0;
-    keypad(win, TRUE);
-
-    while (!done) {
-        werase(win);
-        box(win, 0, 0);
-
-        for (int i = 0; i < FIELD_COUNT; ++i) {
-            if (i == highlight)
-                wattron(win, A_REVERSE);
-
-            switch (i) {
-            case FIELD_ENABLE_COLOR:
-                mvwprintw(win, i + 1, 2, "%s: %s", labels[i],
-                          cfg->enable_color ? "Enabled" : "Disabled");
-                break;
-            case FIELD_MOUSE:
-                mvwprintw(win, i + 1, 2, "%s: %s", labels[i],
-                          cfg->enable_mouse ? "Enabled" : "Disabled");
-                break;
-            case FIELD_BACKGROUND:
-                mvwprintw(win, i + 1, 2, "%s: %s", labels[i],
-                          cfg->background_color);
-                break;
-            case FIELD_KEYWORD:
-                mvwprintw(win, i + 1, 2, "%s: %s", labels[i],
-                          cfg->keyword_color);
-                break;
-            case FIELD_COMMENT:
-                mvwprintw(win, i + 1, 2, "%s: %s", labels[i],
-                          cfg->comment_color);
-                break;
-            case FIELD_STRING:
-                mvwprintw(win, i + 1, 2, "%s: %s", labels[i],
-                          cfg->string_color);
-                break;
-            case FIELD_TYPE:
-                mvwprintw(win, i + 1, 2, "%s: %s", labels[i], cfg->type_color);
-                break;
-            case FIELD_SYMBOL:
-                mvwprintw(win, i + 1, 2, "%s: %s", labels[i],
-                          cfg->symbol_color);
-                break;
-            }
-
-            if (i == highlight)
-                wattroff(win, A_REVERSE);
-        }
-
-        mvwprintw(win, win_height - 2, 2,
-                  "Arrows: move  Enter: change  ESC: done");
-        wrefresh(win);
-
-        ch = wgetch(win);
-        if (ch == KEY_UP) {
-            if (highlight > 0)
-                --highlight;
-        } else if (ch == KEY_DOWN) {
-            if (highlight < FIELD_COUNT - 1)
-                ++highlight;
-        } else if (ch == '\n') {
-            switch (highlight) {
-            case FIELD_ENABLE_COLOR:
-                cfg->enable_color =
-                    select_bool("Enable color", cfg->enable_color, win);
-                break;
-            case FIELD_MOUSE:
-                cfg->enable_mouse =
-                    select_bool("Enable mouse", cfg->enable_mouse, win);
-                break;
-            case FIELD_BACKGROUND: {
-                const char *sel = select_color(cfg->background_color, win);
-                if (sel) {
-                    strncpy(cfg->background_color, sel,
-                            sizeof(cfg->background_color) - 1);
-                    cfg->background_color[sizeof(cfg->background_color) - 1] =
-                        '\0';
-                }
-                break;
-            }
-            case FIELD_KEYWORD: {
-                const char *sel = select_color(cfg->keyword_color, win);
-                if (sel) {
-                    strncpy(cfg->keyword_color, sel,
-                            sizeof(cfg->keyword_color) - 1);
-                    cfg->keyword_color[sizeof(cfg->keyword_color) - 1] = '\0';
-                }
-                break;
-            }
-            case FIELD_COMMENT: {
-                const char *sel = select_color(cfg->comment_color, win);
-                if (sel) {
-                    strncpy(cfg->comment_color, sel,
-                            sizeof(cfg->comment_color) - 1);
-                    cfg->comment_color[sizeof(cfg->comment_color) - 1] = '\0';
-                }
-                break;
-            }
-            case FIELD_STRING: {
-                const char *sel = select_color(cfg->string_color, win);
-                if (sel) {
-                    strncpy(cfg->string_color, sel,
-                            sizeof(cfg->string_color) - 1);
-                    cfg->string_color[sizeof(cfg->string_color) - 1] = '\0';
-                }
-                break;
-            }
-            case FIELD_TYPE: {
-                const char *sel = select_color(cfg->type_color, win);
-                if (sel) {
-                    strncpy(cfg->type_color, sel,
-                            sizeof(cfg->type_color) - 1);
-                    cfg->type_color[sizeof(cfg->type_color) - 1] = '\0';
-                }
-                break;
-            }
-            case FIELD_SYMBOL: {
-                const char *sel = select_color(cfg->symbol_color, win);
-                if (sel) {
-                    strncpy(cfg->symbol_color, sel,
-                            sizeof(cfg->symbol_color) - 1);
-                    cfg->symbol_color[sizeof(cfg->symbol_color) - 1] = '\0';
-                }
-                break;
-            }
-            }
-        } else if (ch == 27) {
-            done = 1;
-        }
-    }
-
-    wclear(win);
-    wrefresh(win);
-    delwin(win);
-    wrefresh(stdscr);
-
-    return memcmp(&original, cfg, sizeof(AppConfig)) != 0;
-}
-
-const char *select_color(const char *current, WINDOW *parent) {
-    static const char *colors[] = {
-        "BLACK", "RED", "GREEN", "YELLOW",
-        "BLUE", "MAGENTA", "CYAN", "WHITE"
-    };
-    const int count = sizeof(colors) / sizeof(colors[0]);
-
-    int highlight = 0;
-    if (current) {
-        for (int i = 0; i < count; ++i) {
-            if (strcasecmp(current, colors[i]) == 0) {
-                highlight = i;
-                break;
-            }
-        }
-    }
-
-    int start = 0;
-    int own = 0;
-    int win_height, win_width;
-    WINDOW *win;
-    if (parent) {
-        int ph, pw;
-        getmaxyx(parent, ph, pw);
-        win_height = ph - 4;
-        win_width = pw - 4;
-        win = create_popup_window(win_height, win_width, parent);
-        if (!win)
-            return NULL;
-        own = 1;
-    } else {
-        win_height = LINES - 4;
-        win_width = COLS - 4;
-        win = create_popup_window(win_height, win_width, NULL);
-        if (!win)
-            return NULL;
-        own = 1;
-    }
-    keypad(win, TRUE);
-
-    while (1) {
-        werase(win);
-        box(win, 0, 0);
-
-        int max_display = win_height - 2; // reserve line for instructions
-        if (highlight < start)
-            start = highlight;
-        if (highlight >= start + max_display)
-            start = highlight - max_display + 1;
-
-        for (int i = 0; i < max_display && i + start < count; ++i) {
-            int idx = i + start;
-            if (idx == highlight)
-                wattron(win, A_REVERSE);
-            mvwprintw(win, i + 1, 1, "%s", colors[idx]);
-            wattroff(win, A_REVERSE);
-        }
-
-        for (int i = count - start; i < max_display; ++i) {
-            mvwprintw(win, i + 1, 1, "%*s", win_width - 2, "");
-        }
-
-        mvwprintw(win, win_height - 1, 1, "Arrows: move  Enter: select  ESC: cancel");
-        wrefresh(win);
-
-        int ch = wgetch(win);
-        if (ch == KEY_UP) {
-            if (highlight > 0)
-                --highlight;
-        } else if (ch == KEY_DOWN) {
-            if (highlight < count - 1)
-                ++highlight;
-        } else if (ch == '\n') {
-            if (own) {
-                werase(win);
-                wrefresh(win);
-                delwin(win);
-                if (parent) {
-                    touchwin(parent);
-                    wrefresh(parent);
-                } else {
-                    wrefresh(stdscr);
-                }
-            }
-            return colors[highlight];
-        } else if (ch == 27) {
-            if (own) {
-                werase(win);
-                wrefresh(win);
-                delwin(win);
-                if (parent) {
-                    touchwin(parent);
-                    wrefresh(parent);
-                } else {
-                    wrefresh(stdscr);
-                }
-            }
-            return NULL;
-        }
-    }
-
-    return NULL;
-}
-
-int select_bool(const char *prompt, int current, WINDOW *parent) {
-    static const char *options[] = {"Disabled", "Enabled"};
-    int highlight = current ? 1 : 0;
-
-    int own = 0;
-    int win_height = 6;
-    int win_width = 20;
-    WINDOW *win;
-    if (parent) {
-        win = create_popup_window(win_height, win_width, parent);
-        if (!win)
-            return current;
-        own = 1;
-    } else {
-        win = create_popup_window(win_height, win_width, NULL);
-        if (!win)
-            return current;
-        own = 1;
-    }
-    keypad(win, TRUE);
-
-    while (1) {
-        werase(win);
-        box(win, 0, 0);
-
-        int row = 1;
-        if (prompt) {
-            mvwprintw(win, row++, 1, "%s", prompt);
-        }
-
-        for (int i = 0; i < 2; ++i) {
-            if (highlight == i)
-                wattron(win, A_REVERSE);
-            mvwprintw(win, row + i, 1, "%s", options[i]);
-            wattroff(win, A_REVERSE);
-        }
-
-        mvwprintw(win, win_height - 1, 1, "Arrows: move  Enter: select  ESC: cancel");
-        wrefresh(win);
-
-        int ch = wgetch(win);
-        if (ch == KEY_UP || ch == KEY_LEFT) {
-            if (highlight > 0)
-                --highlight;
-        } else if (ch == KEY_DOWN || ch == KEY_RIGHT) {
-            if (highlight < 1)
-                ++highlight;
-        } else if (ch == '\n') {
-            if (own) {
-                werase(win);
-                wrefresh(win);
-                delwin(win);
-                if (parent) {
-                    touchwin(parent);
-                    wrefresh(parent);
-                } else {
-                    wrefresh(stdscr);
-                }
-            }
-            return highlight == 1 ? 1 : 0;
-        } else if (ch == 27) {
-            if (own) {
-                werase(win);
-                wrefresh(win);
-                delwin(win);
-                if (parent) {
-                    touchwin(parent);
-                    wrefresh(parent);
-                } else {
-                    wrefresh(stdscr);
-                }
-            }
-            return current;
-        }
-    }
-
-    return current;
-}
-
-
